revert to vanilla js
